% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_gt.R
\name{long_gt}
\alias{long_gt}
\title{Long term Google Trends queries by overlapped windows}
\usage{
long_gt(
  keyword = NULL,
  geo = "",
  input.sdate,
  input.edate,
  input.frequency = "d",
  input.delta = 6,
  input.ol.win = 1,
  input.type = "web",
  input.categ = 0
)
}
\arguments{
\item{keyword}{A character vector with the actual Google Trends query keywords. Multiple
keywords are NOT possible right now.}

\item{geo}{A character vector denoting geographic regions for the query, default to
all for global queries. Multiple regions are NOT possible right now.}

\item{input.sdate}{Starting date for the query search. Should be Date format "Y-m-d".}

\item{input.edate}{Ending date for the query search. Should be Date format "Y-m-d".}

\item{input.frequency}{Frequency of the resulting query. Valid options are:
\itemize{
\item "d" - Daily (default).
\item "w" - Weekly.
\item "m" - Monthly.
\item "q" - Quarterly.
}}

\item{input.delta}{A nonzero positive integer indicating the window length of the query in months
(if input.frequency="d") or years (if input.frequency="w").
Notice that it will be not used in another case.}

\item{input.ol.win}{A nonzero positive integer indicating the overlapping window length.
Default is 1 (month/year).}

\item{input.type}{A character string defining the Google product for which the trend
query if preformed. Valid options are: "web" (default), "news",
"images", "froogle", "youtube".}

\item{input.categ}{A character denoting the category, defaults to “0”.}
}
\value{
An object of class list containing two objects: A data.frame object with
columns date, Google Trends indicator and an overlapped period dummy (if
daily or weekly query is performed). A ggplot figure of the resulting time
series with shaded areas showing overlapping periods.
}
\description{
The long_gt method performs Google Trends queries for the 'keyword' argument for
the period spanned between starting and ending date by overlapping windows. The function
purpose is to produce long term high frequency Google Trends series by overlapping
queries on the same 'keyword' but different time span. Regular queries of short
duration periods could also be performed.
}
\details{
The function purpose is to produce long term high frequency Google Trends
series by overlapping queries of the same 'keyword' but with different time span.
Regular queries of short periods could also be performed. The function is time consuming
due to the need to perform different queries, and a random sleep time between 2 and 5
seconds for each query step in order to prevent that Google blocks the IP.
}
\examples{
# Search for daily data on word hotel in Colombia from 2014 to 2021
 res <- long_gt(keyword = "hotel",geo="CO",
                input.sdate = as.Date("2014-01-01"),
                input.edate = as.Date("2021-04-16"),
                input.frequency = "d",input.delta = 6,
                input.ol.win = 1,input.type="web",input.categ=0)
}
\seealso{
long_gt_ltc
}
\author{
Sebastian Sanin-Restrepo
}
