% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_gt_ltc.R
\name{long_gt_ltc}
\alias{long_gt_ltc}
\title{Corrected by long-term trend - Long term Google Trends queries by overlapped windows}
\usage{
long_gt_ltc(
  keyword = NULL,
  geo = "",
  input.sdate,
  input.edate,
  input.frequency = "d",
  td.method = "chow-lin",
  input.delta = 6,
  input.ol.win = 1,
  input.type = "web",
  input.categ = 0
)
}
\arguments{
\item{keyword}{A character vector with the actual Google Trends query keywords. Multiple
keywords are NOT possible right now.}

\item{geo}{A character vector denoting geographic regions for the query, default to
all for global queries. Multiple regions are NOT possible right now.}

\item{input.sdate}{Starting date for the query search. Should be Date format "Y-m-d".}

\item{input.edate}{Ending date for the query search. Should be Date format "Y-m-d".}

\item{input.frequency}{Frequency of the resulting query. Valid options are:
\itemize{
\item "d" - Daily (default).
\item "w" - Weekly.
\item "m" - Monthly.
\item "q" - Quarterly.
}}

\item{td.method}{Character indicating method for temporal disaggregation. "denton" refers
to "Denton-Cholette" and "chow-lin" (default) referes to "chow-lin maxlog".}

\item{input.delta}{An integer indicating the window length of the query in months
(if input.frequency="d") or years (if input.frequency="w").
Notice that it will be not used in another case.}

\item{input.ol.win}{An integer indicating the overlapping window length.
Default is 1 (month/year).}

\item{input.type}{A character string defining the Google product for which the trend
query if preformed. Valid options are: "web" (default), "news",
"images", "froogle", "youtube".}

\item{input.categ}{A character denoting the category, defaults to “0”.}
}
\value{
An object of class list containing two objects: A data.frame object with
columns date, Google Trends indicator and an overlapped period dummy (if
daily or weekly query is performed). A ggplot figure of the resulting time
series with shaded areas showing overlapping periods and monthly trend.
}
\description{
The long_gt_ltc method performs Google Trends queries for the 'keyword'
argument for the period spanned between starting and ending date by overlapping windows.
It also performs a correction for long-term query trend by using temporal disaggregation
of monthly query using the high frequency query as indicator.
}
\details{
The function purpose is to produce long term high frequency Google Trends
series by overlapping queries of the same 'keyword' but with different time span. It also
perform a correction for long-term monthly trend by applying temporal disaggregation methods
(Denton-Cholette and Chow-Lin Maxlog) using high frequency overlapped index as indicator.
Regular queries of short periods could also be performed. The function is time consuming
due to the need to perform different queries, and a random sleep time between 2 and 5
seconds for each query step in order to prevent that Google blocks the IP.
}
\examples{
# Search for daily data on word hotel in Colombia from 2014 to 2021
 res <- long_gt_ltc(keyword = "hotel",geo="CO",
                input.sdate = as.Date("2014-01-01"),
                input.edate = as.Date("2021-04-16"),
                input.frequency = "d",td.method="chow-lin",
                input.delta = 6,
                input.ol.win = 1)
}
\seealso{
long_gt_ltc
}
\author{
Sebastian Sanin-Restrepo
}
