% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninTrenDo.R
\docType{package}
\name{ninTrenDo}
\alias{ninTrenDo}
\title{ninTrenDo: Perform long-term high frequency Google Trends queries by overlapping}
\description{
Reconstruct Google Trends high frequency data for extended periods. Using
overlapping windows and letting for long-term trend correction, this
package allows the search for any single query of daily, weekly, monthly
or quarterly frequency, covering longer periods than those allowed by
Google Trends direct search.
}
\section{ninTrenDo functions}{

\itemize{
\item long_gt: Long term Google Trends queries by overlapped windows
\item long_gt_ltc: Corrected by long-term trend - Long term Google Trends
queries by overlapped windows.
}
}

